Starting the CSS cleanup
TODO: Don’t have all the styles for the message list and such yet.

The goal of this rewrite is to:

(1) reduce CSS code by only using what we need and reusing classes as
much as possible. We should almost never be styling based on id.

(2) using it as a way to clean up the HTML to simplify. e.g., how the
“action-btn” is now simpler than it used to be (and a reusable
component we can drop in.