fix(modify-participant) make patch work on participant update form.
This change was hard because the serialized participant's values were
different from the saved database and therefor the crispy form was not
rendering with the current participants values.

The fields to edit in the update form are now have a _display value as
well in the participant serializer.

Doing this made the NullBoolean on hiv_disclosed hard so I changed it to
a CharField with choices.